API/requirements.txt
psycopg2-binary
Flask-Caching
threaded
flask
redis
pika
uuid
